Testing framework review fixes
