--- integtest/src/main/java/org/corfudb/universe/cluster/vm/VmCluster.java
@@ -1,38 +1,0 @@
-package org.corfudb.universe.cluster.vm;
-
-import com.google.common.collect.ImmutableMap;
-import org.corfudb.universe.cluster.Cluster;
-import org.corfudb.universe.cluster.ClusterException;
-import org.corfudb.universe.service.Service;
-
-/**
- * This implementation provides a Cluster servers each of which a VM machine.
- * Deploy is going to deploy a list of servers
- */
-public class VmCluster implements Cluster {
-
-    @Override
-    public Cluster deploy() throws ClusterException {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public void shutdown() throws ClusterException {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public ClusterParams getClusterParams() {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public ImmutableMap<String, Service> services() {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-
-    @Override
-    public Service getService(String serviceName) {
-        throw new UnsupportedOperationException("Not implemented");
-    }
-}
