+++ it/doc/testing-framework-design.png
