+++ it/src/test/resources/scenarios/scenario.json
